# docs/openapi.yaml
openapi: 3.0.0
info:
  title: API Turnos
  description: 'Documentación para API Turnos'
  contact:
    email: gonza.fede.sanchez@gmail.com
  version: 1.0.0
servers:
- url: localhost
paths:
...
  /Clinica:
    get:
      tags:
      - Clínica
      summary: Obtener un listado de las Clínicas
      operationId: obtenerClinica
      responses:
        200:
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Clinica'
        500:
          description: Hubo un error al obtener el listado de las Clínicas
          content: {}
    post:
      tags:
      - Clínica
      summary: Agregar una nueva Clínica
      operationId: agregarClinica
      requestBody:
        description: Objeto Clínica a ser creado
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Clinica'
        required: true
      responses:
        201:
          description: La Historia Clínica fue creada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Clinica'
        400:
          description: Los datos de la Clínica son incorrectos
          content: {}
        500:
          description: Ocurrió un error al crear la Clínica
          content: {}
  /Clinica/{id}:    
    delete:
      tags: 
        - Clínica
      summary: Eliminar una Clínica especifica
      operationId: eliminarClinica
      parameters: 
        - in: path
          name: id
          required: true
          schema:
            type: string 
            example: "b11faa60-f287-11e9-8a31-dd8e324cb24b"
      responses:
        400:
          description: Operación Exitosa
          content: {}
        500:
          description: Hubo un error al eliminar la Clínica
          content: {}  
    patch:
      tags:
        - Clínica
      summary: Modificar datos de una Clínica por su id
      operationId: actualizarClinica
      parameters: 
      - in: path
        name: id
        required: true
        schema:
          type: string
          example: "b11faa60-f287-11e9-8a31-dd8e324cb24b"
      responses:
        200:
          description: Operacion Exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Clinica'
        500:
          description: Hubo un error al modificar la Clínica
          content: {}              
components:
  schemas:
...
    Clinica:
      required:
      - cuit
      - nombre
      - tel
      - dir
      type: object
      properties:
        _id:
          type: string
          example: b11faa60-f287-11e9-8a31-dd8e324cb24b
        cuit:
          type: string
          example: 30-30587989-5
        nombre:
          type: string
          example: Clinica Mercedario
        tel:
          type: int
          example: 4237852 
        dir:
          type: string
          example: España 140 N